Using md5sum to make upload faster.
Instead of replacing all files during the build, each output file
is first built and then compared with the previous version.  if
there is a change then the old file is replaced, otherwise it is
kept.

This trick allows to keep file attributes like last modification
date, and ends up in optimizing the RSync upload.